+++ App.js
@@ -1,30 +1,18 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import Button from 'react-native-button';
-
-import NavBar from './components/NavBar'
+import { StackNavigator } from 'react-navigation';
+// import NavBar from './components/NavBar'
 import Home from './components/Home'
-
-export default class App extends React.Component {
-
-  render() {
-    return (
-      <View style={styles.container}>
-
-        <NavBar/>
-
-        <Home />
-
-
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-  },
+// import App from './components/App'
+// import HomeButton from './components/HomeButton'
+// const MainScreen = ({ navigation }) => (
+//   <View>
+//   <Text>Hello</Text>
+//   </View>
+// )
+const AppRoutes = StackNavigator({
+  Home: {screen: Home}
 });
+
+export default () => <AppRoutes />
