+++ components/HomeButton.js
@@ -10,11 +10,12 @@ export default class HomeButton extends Component {
   }
   render() {
     return (
-      <Button
-        containerStyle={{padding:10, height:40, borderRadius:4, backgroundColor: 'green', flex: .5, marginBottom: 10}}
-        style={{fontSize: 20, color: 'white',}}>
+      <Button>
           Create a New Report
       </Button>
     );
   }
 };
+
+// containerStyle={{padding:10, height:40, borderRadius:4, backgroundColor: 'green', flex: .5, marginBottom: 10}}
+// style={{fontSize: 20, color: 'white',}}>
