+++ package.json
@@ -22,6 +22,7 @@
     "expo": "^16.0.0",
     "react": "16.0.0-alpha.6",
     "react-native": "^0.43.4",
-    "react-native-button": "^1.8.2"
+    "react-native-button": "^1.8.2",
+    "react-navigation": "^1.0.0-beta.11"
   }
 }
